refactor(core): Split resolveEffect.ts into category-based modules for LLM optimization
Split the 875-line resolveEffect.ts into category-based modules following
the pattern established in packages/shared/src/events/:

New modules:
- types.ts: EffectResolutionResult type definition
- resolvability.ts: isEffectResolvable and canObtainBasicColorMana
- compound.ts: CompoundEffect, ConditionalEffect, ScalingEffect handlers
- choice.ts: ChoiceEffect handler
- crystallize.ts: ConvertManaToCrystal, CrystallizeColor handlers
- cardBoostResolvers.ts: CardBoost, ResolveBoostTarget handlers
- combatEffects.ts: SelectCombatEnemy, ResolveCombatEnemyTarget handlers
- reverse.ts: reverseEffect function for undo

Updated index.ts to:
- Re-export all modules with comprehensive documentation
- Contain the main resolveEffect dispatcher
- Route to category-specific handlers

Each category module has:
- Header documentation with JSDoc module description
- Section separators for visual organization
- Well-documented exported functions with @param, @returns, @example
- Clear imports using .js extensions for ESM compatibility

Updated imports in dependent files:
- commands/playCardCommand.ts
- commands/resolveChoiceCommand.ts
- validActions/cards.ts
- effectUndoContext.ts
- atomicEffects.ts, manaDrawEffects.ts, unitEffects.ts
- All test files

No behavior changes - purely structural refactor to improve LLM
comprehension by reducing file size and grouping related logic.
