+++ packages/core/src/engine/__tests__/conditionalEffects.test.ts
@@ -5,7 +5,7 @@
 import { describe, it, expect } from "vitest";
 import { createTestGameState, createTestPlayer, createTestHex } from "./testHelpers.js";
 import { evaluateCondition } from "../effects/conditionEvaluator.js";
-import { resolveEffect } from "../effects/resolveEffect.js";
+import { resolveEffect } from "../effects/index.js";
 import {
   CONDITION_TIME_OF_DAY,
   CONDITION_IN_PHASE,
